Split credit-action panel into its own component
Decouple CreditActionPanel from UserProfilePanel so each can be placed independently; compose both in RightColumn.
